+++ lib/health.g.dart
@@ -6,18 +6,25 @@ part of 'health.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-HealthDataPoint _$HealthDataPointFromJson(Map<String, dynamic> json) => HealthDataPoint(
+HealthDataPoint _$HealthDataPointFromJson(
+  Map<String, dynamic> json,
+) => HealthDataPoint(
   uuid: json['uuid'] as String,
   value: HealthValue.fromJson(json['value'] as Map<String, dynamic>),
   type: $enumDecode(_$HealthDataTypeEnumMap, json['type']),
   unit: $enumDecode(_$HealthDataUnitEnumMap, json['unit']),
   dateFrom: DateTime.parse(json['dateFrom'] as String),
   dateTo: DateTime.parse(json['dateTo'] as String),
-  sourcePlatform: $enumDecode(_$HealthPlatformTypeEnumMap, json['sourcePlatform']),
+  sourcePlatform: $enumDecode(
+    _$HealthPlatformTypeEnumMap,
+    json['sourcePlatform'],
+  ),
   sourceDeviceId: json['sourceDeviceId'] as String,
   sourceId: json['sourceId'] as String,
   sourceName: json['sourceName'] as String,
-  recordingMethod: $enumDecodeNullable(_$RecordingMethodEnumMap, json['recordingMethod']) ?? RecordingMethod.unknown,
+  recordingMethod:
+      $enumDecodeNullable(_$RecordingMethodEnumMap, json['recordingMethod']) ??
+      RecordingMethod.unknown,
   workoutSummary: json['workoutSummary'] == null
       ? null
       : WorkoutSummary.fromJson(json['workoutSummary'] as Map<String, dynamic>),
@@ -25,22 +32,23 @@ HealthDataPoint _$HealthDataPointFromJson(Map<String, dynamic> json) => HealthDa
   deviceModel: json['deviceModel'] as String?,
 );
 
-Map<String, dynamic> _$HealthDataPointToJson(HealthDataPoint instance) => <String, dynamic>{
-  'uuid': instance.uuid,
-  'value': instance.value.toJson(),
-  'type': _$HealthDataTypeEnumMap[instance.type]!,
-  'unit': _$HealthDataUnitEnumMap[instance.unit]!,
-  'dateFrom': instance.dateFrom.toIso8601String(),
-  'dateTo': instance.dateTo.toIso8601String(),
-  'sourcePlatform': _$HealthPlatformTypeEnumMap[instance.sourcePlatform]!,
-  'sourceDeviceId': instance.sourceDeviceId,
-  'sourceId': instance.sourceId,
-  'sourceName': instance.sourceName,
-  'recordingMethod': _$RecordingMethodEnumMap[instance.recordingMethod]!,
-  'workoutSummary': ?instance.workoutSummary?.toJson(),
-  'metadata': ?instance.metadata,
-  'deviceModel': ?instance.deviceModel,
-};
+Map<String, dynamic> _$HealthDataPointToJson(HealthDataPoint instance) =>
+    <String, dynamic>{
+      'uuid': instance.uuid,
+      'value': instance.value.toJson(),
+      'type': _$HealthDataTypeEnumMap[instance.type]!,
+      'unit': _$HealthDataUnitEnumMap[instance.unit]!,
+      'dateFrom': instance.dateFrom.toIso8601String(),
+      'dateTo': instance.dateTo.toIso8601String(),
+      'sourcePlatform': _$HealthPlatformTypeEnumMap[instance.sourcePlatform]!,
+      'sourceDeviceId': instance.sourceDeviceId,
+      'sourceId': instance.sourceId,
+      'sourceName': instance.sourceName,
+      'recordingMethod': _$RecordingMethodEnumMap[instance.recordingMethod]!,
+      'workoutSummary': ?instance.workoutSummary?.toJson(),
+      'metadata': ?instance.metadata,
+      'deviceModel': ?instance.deviceModel,
+    };
 
 const _$HealthDataTypeEnumMap = {
   HealthDataType.ACTIVE_ENERGY_BURNED: 'ACTIVE_ENERGY_BURNED',
@@ -151,6 +159,7 @@ const _$HealthDataTypeEnumMap = {
   HealthDataType.ELECTRODERMAL_ACTIVITY: 'ELECTRODERMAL_ACTIVITY',
   HealthDataType.ELECTROCARDIOGRAM: 'ELECTROCARDIOGRAM',
   HealthDataType.TOTAL_CALORIES_BURNED: 'TOTAL_CALORIES_BURNED',
+  HealthDataType.ACTIVITY_INTENSITY: 'ACTIVITY_INTENSITY',
 };
 
 const _$HealthDataUnitEnumMap = {
@@ -178,7 +187,8 @@ const _$HealthDataUnitEnumMap = {
   HealthDataUnit.INCHES_OF_MERCURY: 'INCHES_OF_MERCURY',
   HealthDataUnit.CENTIMETER_OF_WATER: 'CENTIMETER_OF_WATER',
   HealthDataUnit.ATMOSPHERE: 'ATMOSPHERE',
-  HealthDataUnit.DECIBEL_A_WEIGHTED_SOUND_PRESSURE_LEVEL: 'DECIBEL_A_WEIGHTED_SOUND_PRESSURE_LEVEL',
+  HealthDataUnit.DECIBEL_A_WEIGHTED_SOUND_PRESSURE_LEVEL:
+      'DECIBEL_A_WEIGHTED_SOUND_PRESSURE_LEVEL',
   HealthDataUnit.SECOND: 'SECOND',
   HealthDataUnit.MILLISECOND: 'MILLISECOND',
   HealthDataUnit.MINUTE: 'MINUTE',
@@ -219,51 +229,81 @@ const _$RecordingMethodEnumMap = {
   RecordingMethod.manual: 'manual',
 };
 
-HealthValue _$HealthValueFromJson(Map<String, dynamic> json) => HealthValue()..$type = json['__type'] as String?;
+HealthValue _$HealthValueFromJson(Map<String, dynamic> json) =>
+    HealthValue()..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$HealthValueToJson(HealthValue instance) => <String, dynamic>{'__type': ?instance.$type};
+Map<String, dynamic> _$HealthValueToJson(HealthValue instance) =>
+    <String, dynamic>{'__type': ?instance.$type};
 
 NumericHealthValue _$NumericHealthValueFromJson(Map<String, dynamic> json) =>
-    NumericHealthValue(numericValue: json['numericValue'] as num)..$type = json['__type'] as String?;
+    NumericHealthValue(numericValue: json['numericValue'] as num)
+      ..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$NumericHealthValueToJson(NumericHealthValue instance) => <String, dynamic>{
-  '__type': ?instance.$type,
-  'numericValue': instance.numericValue,
-};
+Map<String, dynamic> _$NumericHealthValueToJson(NumericHealthValue instance) =>
+    <String, dynamic>{
+      '__type': ?instance.$type,
+      'numericValue': instance.numericValue,
+    };
 
-AudiogramHealthValue _$AudiogramHealthValueFromJson(Map<String, dynamic> json) => AudiogramHealthValue(
-  frequencies: (json['frequencies'] as List<dynamic>).map((e) => e as num).toList(),
-  leftEarSensitivities: (json['leftEarSensitivities'] as List<dynamic>).map((e) => e as num).toList(),
-  rightEarSensitivities: (json['rightEarSensitivities'] as List<dynamic>).map((e) => e as num).toList(),
+AudiogramHealthValue _$AudiogramHealthValueFromJson(
+  Map<String, dynamic> json,
+) => AudiogramHealthValue(
+  frequencies: (json['frequencies'] as List<dynamic>)
+      .map((e) => e as num)
+      .toList(),
+  leftEarSensitivities: (json['leftEarSensitivities'] as List<dynamic>)
+      .map((e) => e as num)
+      .toList(),
+  rightEarSensitivities: (json['rightEarSensitivities'] as List<dynamic>)
+      .map((e) => e as num)
+      .toList(),
 )..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$AudiogramHealthValueToJson(AudiogramHealthValue instance) => <String, dynamic>{
+Map<String, dynamic> _$AudiogramHealthValueToJson(
+  AudiogramHealthValue instance,
+) => <String, dynamic>{
   '__type': ?instance.$type,
   'frequencies': instance.frequencies,
   'leftEarSensitivities': instance.leftEarSensitivities,
   'rightEarSensitivities': instance.rightEarSensitivities,
 };
 
-WorkoutHealthValue _$WorkoutHealthValueFromJson(Map<String, dynamic> json) => WorkoutHealthValue(
-  workoutActivityType: $enumDecode(_$HealthWorkoutActivityTypeEnumMap, json['workoutActivityType']),
-  totalEnergyBurned: (json['totalEnergyBurned'] as num?)?.toInt(),
-  totalEnergyBurnedUnit: $enumDecodeNullable(_$HealthDataUnitEnumMap, json['totalEnergyBurnedUnit']),
-  totalDistance: (json['totalDistance'] as num?)?.toInt(),
-  totalDistanceUnit: $enumDecodeNullable(_$HealthDataUnitEnumMap, json['totalDistanceUnit']),
-  totalSteps: (json['totalSteps'] as num?)?.toInt(),
-  totalStepsUnit: $enumDecodeNullable(_$HealthDataUnitEnumMap, json['totalStepsUnit']),
-)..$type = json['__type'] as String?;
+WorkoutHealthValue _$WorkoutHealthValueFromJson(Map<String, dynamic> json) =>
+    WorkoutHealthValue(
+      workoutActivityType: $enumDecode(
+        _$HealthWorkoutActivityTypeEnumMap,
+        json['workoutActivityType'],
+      ),
+      totalEnergyBurned: (json['totalEnergyBurned'] as num?)?.toInt(),
+      totalEnergyBurnedUnit: $enumDecodeNullable(
+        _$HealthDataUnitEnumMap,
+        json['totalEnergyBurnedUnit'],
+      ),
+      totalDistance: (json['totalDistance'] as num?)?.toInt(),
+      totalDistanceUnit: $enumDecodeNullable(
+        _$HealthDataUnitEnumMap,
+        json['totalDistanceUnit'],
+      ),
+      totalSteps: (json['totalSteps'] as num?)?.toInt(),
+      totalStepsUnit: $enumDecodeNullable(
+        _$HealthDataUnitEnumMap,
+        json['totalStepsUnit'],
+      ),
+    )..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$WorkoutHealthValueToJson(WorkoutHealthValue instance) => <String, dynamic>{
-  '__type': ?instance.$type,
-  'workoutActivityType': _$HealthWorkoutActivityTypeEnumMap[instance.workoutActivityType]!,
-  'totalEnergyBurned': ?instance.totalEnergyBurned,
-  'totalEnergyBurnedUnit': ?_$HealthDataUnitEnumMap[instance.totalEnergyBurnedUnit],
-  'totalDistance': ?instance.totalDistance,
-  'totalDistanceUnit': ?_$HealthDataUnitEnumMap[instance.totalDistanceUnit],
-  'totalSteps': ?instance.totalSteps,
-  'totalStepsUnit': ?_$HealthDataUnitEnumMap[instance.totalStepsUnit],
-};
+Map<String, dynamic> _$WorkoutHealthValueToJson(WorkoutHealthValue instance) =>
+    <String, dynamic>{
+      '__type': ?instance.$type,
+      'workoutActivityType':
+          _$HealthWorkoutActivityTypeEnumMap[instance.workoutActivityType]!,
+      'totalEnergyBurned': ?instance.totalEnergyBurned,
+      'totalEnergyBurnedUnit':
+          ?_$HealthDataUnitEnumMap[instance.totalEnergyBurnedUnit],
+      'totalDistance': ?instance.totalDistance,
+      'totalDistanceUnit': ?_$HealthDataUnitEnumMap[instance.totalDistanceUnit],
+      'totalSteps': ?instance.totalSteps,
+      'totalStepsUnit': ?_$HealthDataUnitEnumMap[instance.totalStepsUnit],
+    };
 
 const _$HealthWorkoutActivityTypeEnumMap = {
   HealthWorkoutActivityType.AMERICAN_FOOTBALL: 'AMERICAN_FOOTBALL',
@@ -284,7 +324,8 @@ const _$HealthWorkoutActivityTypeEnumMap = {
   HealthWorkoutActivityType.GOLF: 'GOLF',
   HealthWorkoutActivityType.GYMNASTICS: 'GYMNASTICS',
   HealthWorkoutActivityType.HANDBALL: 'HANDBALL',
-  HealthWorkoutActivityType.HIGH_INTENSITY_INTERVAL_TRAINING: 'HIGH_INTENSITY_INTERVAL_TRAINING',
+  HealthWorkoutActivityType.HIGH_INTENSITY_INTERVAL_TRAINING:
+      'HIGH_INTENSITY_INTERVAL_TRAINING',
   HealthWorkoutActivityType.HIKING: 'HIKING',
   HealthWorkoutActivityType.HOCKEY: 'HOCKEY',
   HealthWorkoutActivityType.JUMP_ROPE: 'JUMP_ROPE',
@@ -320,7 +361,8 @@ const _$HealthWorkoutActivityTypeEnumMap = {
   HealthWorkoutActivityType.FISHING: 'FISHING',
   HealthWorkoutActivityType.FITNESS_GAMING: 'FITNESS_GAMING',
   HealthWorkoutActivityType.FLEXIBILITY: 'FLEXIBILITY',
-  HealthWorkoutActivityType.FUNCTIONAL_STRENGTH_TRAINING: 'FUNCTIONAL_STRENGTH_TRAINING',
+  HealthWorkoutActivityType.FUNCTIONAL_STRENGTH_TRAINING:
+      'FUNCTIONAL_STRENGTH_TRAINING',
   HealthWorkoutActivityType.HAND_CYCLING: 'HAND_CYCLING',
   HealthWorkoutActivityType.HUNTING: 'HUNTING',
   HealthWorkoutActivityType.LACROSSE: 'LACROSSE',
@@ -329,7 +371,8 @@ const _$HealthWorkoutActivityTypeEnumMap = {
   HealthWorkoutActivityType.PADDLE_SPORTS: 'PADDLE_SPORTS',
   HealthWorkoutActivityType.PICKLEBALL: 'PICKLEBALL',
   HealthWorkoutActivityType.PLAY: 'PLAY',
-  HealthWorkoutActivityType.PREPARATION_AND_RECOVERY: 'PREPARATION_AND_RECOVERY',
+  HealthWorkoutActivityType.PREPARATION_AND_RECOVERY:
+      'PREPARATION_AND_RECOVERY',
   HealthWorkoutActivityType.SNOW_SPORTS: 'SNOW_SPORTS',
   HealthWorkoutActivityType.SOCIAL_DANCE: 'SOCIAL_DANCE',
   HealthWorkoutActivityType.STAIRS: 'STAIRS',
@@ -337,7 +380,8 @@ const _$HealthWorkoutActivityTypeEnumMap = {
   HealthWorkoutActivityType.SURFING: 'SURFING',
   HealthWorkoutActivityType.TAI_CHI: 'TAI_CHI',
   HealthWorkoutActivityType.TRACK_AND_FIELD: 'TRACK_AND_FIELD',
-  HealthWorkoutActivityType.TRADITIONAL_STRENGTH_TRAINING: 'TRADITIONAL_STRENGTH_TRAINING',
+  HealthWorkoutActivityType.TRADITIONAL_STRENGTH_TRAINING:
+      'TRADITIONAL_STRENGTH_TRAINING',
   HealthWorkoutActivityType.WATER_FITNESS: 'WATER_FITNESS',
   HealthWorkoutActivityType.WATER_SPORTS: 'WATER_SPORTS',
   HealthWorkoutActivityType.WHEELCHAIR_RUN_PACE: 'WHEELCHAIR_RUN_PACE',
@@ -367,7 +411,9 @@ const _$HealthWorkoutActivityTypeEnumMap = {
   HealthWorkoutActivityType.OTHER: 'OTHER',
 };
 
-WorkoutRouteLocation _$WorkoutRouteLocationFromJson(Map<String, dynamic> json) => WorkoutRouteLocation(
+WorkoutRouteLocation _$WorkoutRouteLocationFromJson(
+  Map<String, dynamic> json,
+) => WorkoutRouteLocation(
   latitude: (json['latitude'] as num).toDouble(),
   longitude: (json['longitude'] as num).toDouble(),
   timestamp: DateTime.parse(json['timestamp'] as String),
@@ -380,7 +426,9 @@ WorkoutRouteLocation _$WorkoutRouteLocationFromJson(Map<String, dynamic> json) =
   courseAccuracy: (json['courseAccuracy'] as num?)?.toDouble(),
 )..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$WorkoutRouteLocationToJson(WorkoutRouteLocation instance) => <String, dynamic>{
+Map<String, dynamic> _$WorkoutRouteLocationToJson(
+  WorkoutRouteLocation instance,
+) => <String, dynamic>{
   '__type': ?instance.$type,
   'latitude': instance.latitude,
   'longitude': instance.longitude,
@@ -394,67 +442,90 @@ Map<String, dynamic> _$WorkoutRouteLocationToJson(WorkoutRouteLocation instance)
   'courseAccuracy': ?instance.courseAccuracy,
 };
 
-WorkoutRouteHealthValue _$WorkoutRouteHealthValueFromJson(Map<String, dynamic> json) => WorkoutRouteHealthValue(
+WorkoutRouteHealthValue _$WorkoutRouteHealthValueFromJson(
+  Map<String, dynamic> json,
+) => WorkoutRouteHealthValue(
   locations: (json['locations'] as List<dynamic>)
       .map((e) => WorkoutRouteLocation.fromJson(e as Map<String, dynamic>))
       .toList(),
   workoutUuid: json['workoutUuid'] as String?,
 )..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$WorkoutRouteHealthValueToJson(WorkoutRouteHealthValue instance) => <String, dynamic>{
+Map<String, dynamic> _$WorkoutRouteHealthValueToJson(
+  WorkoutRouteHealthValue instance,
+) => <String, dynamic>{
   '__type': ?instance.$type,
   'locations': instance.locations.map((e) => e.toJson()).toList(),
   'workoutUuid': ?instance.workoutUuid,
 };
 
-ElectrocardiogramHealthValue _$ElectrocardiogramHealthValueFromJson(Map<String, dynamic> json) =>
-    ElectrocardiogramHealthValue(
-      voltageValues: (json['voltageValues'] as List<dynamic>)
-          .map((e) => ElectrocardiogramVoltageValue.fromJson(e as Map<String, dynamic>))
-          .toList(),
-      averageHeartRate: json['averageHeartRate'] as num?,
-      samplingFrequency: (json['samplingFrequency'] as num?)?.toDouble(),
-      classification: $enumDecodeNullable(_$ElectrocardiogramClassificationEnumMap, json['classification']),
-    )..$type = json['__type'] as String?;
+ElectrocardiogramHealthValue _$ElectrocardiogramHealthValueFromJson(
+  Map<String, dynamic> json,
+) => ElectrocardiogramHealthValue(
+  voltageValues: (json['voltageValues'] as List<dynamic>)
+      .map(
+        (e) =>
+            ElectrocardiogramVoltageValue.fromJson(e as Map<String, dynamic>),
+      )
+      .toList(),
+  averageHeartRate: json['averageHeartRate'] as num?,
+  samplingFrequency: (json['samplingFrequency'] as num?)?.toDouble(),
+  classification: $enumDecodeNullable(
+    _$ElectrocardiogramClassificationEnumMap,
+    json['classification'],
+  ),
+)..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$ElectrocardiogramHealthValueToJson(ElectrocardiogramHealthValue instance) => <String, dynamic>{
+Map<String, dynamic> _$ElectrocardiogramHealthValueToJson(
+  ElectrocardiogramHealthValue instance,
+) => <String, dynamic>{
   '__type': ?instance.$type,
   'voltageValues': instance.voltageValues.map((e) => e.toJson()).toList(),
   'averageHeartRate': ?instance.averageHeartRate,
   'samplingFrequency': ?instance.samplingFrequency,
-  'classification': ?_$ElectrocardiogramClassificationEnumMap[instance.classification],
+  'classification':
+      ?_$ElectrocardiogramClassificationEnumMap[instance.classification],
 };
 
 const _$ElectrocardiogramClassificationEnumMap = {
   ElectrocardiogramClassification.NOT_SET: 'NOT_SET',
   ElectrocardiogramClassification.SINUS_RHYTHM: 'SINUS_RHYTHM',
   ElectrocardiogramClassification.ATRIAL_FIBRILLATION: 'ATRIAL_FIBRILLATION',
-  ElectrocardiogramClassification.INCONCLUSIVE_LOW_HEART_RATE: 'INCONCLUSIVE_LOW_HEART_RATE',
-  ElectrocardiogramClassification.INCONCLUSIVE_HIGH_HEART_RATE: 'INCONCLUSIVE_HIGH_HEART_RATE',
-  ElectrocardiogramClassification.INCONCLUSIVE_POOR_READING: 'INCONCLUSIVE_POOR_READING',
+  ElectrocardiogramClassification.INCONCLUSIVE_LOW_HEART_RATE:
+      'INCONCLUSIVE_LOW_HEART_RATE',
+  ElectrocardiogramClassification.INCONCLUSIVE_HIGH_HEART_RATE:
+      'INCONCLUSIVE_HIGH_HEART_RATE',
+  ElectrocardiogramClassification.INCONCLUSIVE_POOR_READING:
+      'INCONCLUSIVE_POOR_READING',
   ElectrocardiogramClassification.INCONCLUSIVE_OTHER: 'INCONCLUSIVE_OTHER',
   ElectrocardiogramClassification.UNRECOGNIZED: 'UNRECOGNIZED',
 };
 
-ElectrocardiogramVoltageValue _$ElectrocardiogramVoltageValueFromJson(Map<String, dynamic> json) =>
-    ElectrocardiogramVoltageValue(
-      voltage: json['voltage'] as num,
-      timeSinceSampleStart: json['timeSinceSampleStart'] as num,
-    )..$type = json['__type'] as String?;
+ElectrocardiogramVoltageValue _$ElectrocardiogramVoltageValueFromJson(
+  Map<String, dynamic> json,
+) => ElectrocardiogramVoltageValue(
+  voltage: json['voltage'] as num,
+  timeSinceSampleStart: json['timeSinceSampleStart'] as num,
+)..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$ElectrocardiogramVoltageValueToJson(ElectrocardiogramVoltageValue instance) => <String, dynamic>{
+Map<String, dynamic> _$ElectrocardiogramVoltageValueToJson(
+  ElectrocardiogramVoltageValue instance,
+) => <String, dynamic>{
   '__type': ?instance.$type,
   'voltage': instance.voltage,
   'timeSinceSampleStart': instance.timeSinceSampleStart,
 };
 
-InsulinDeliveryHealthValue _$InsulinDeliveryHealthValueFromJson(Map<String, dynamic> json) =>
-    InsulinDeliveryHealthValue(
-      units: (json['units'] as num).toDouble(),
-      reason: $enumDecode(_$InsulinDeliveryReasonEnumMap, json['reason']),
-    )..$type = json['__type'] as String?;
+InsulinDeliveryHealthValue _$InsulinDeliveryHealthValueFromJson(
+  Map<String, dynamic> json,
+) => InsulinDeliveryHealthValue(
+  units: (json['units'] as num).toDouble(),
+  reason: $enumDecode(_$InsulinDeliveryReasonEnumMap, json['reason']),
+)..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$InsulinDeliveryHealthValueToJson(InsulinDeliveryHealthValue instance) => <String, dynamic>{
+Map<String, dynamic> _$InsulinDeliveryHealthValueToJson(
+  InsulinDeliveryHealthValue instance,
+) => <String, dynamic>{
   '__type': ?instance.$type,
   'units': instance.units,
   'reason': _$InsulinDeliveryReasonEnumMap[instance.reason]!,
@@ -466,7 +537,9 @@ const _$InsulinDeliveryReasonEnumMap = {
   InsulinDeliveryReason.BOLUS: 'BOLUS',
 };
 
-NutritionHealthValue _$NutritionHealthValueFromJson(Map<String, dynamic> json) => NutritionHealthValue(
+NutritionHealthValue _$NutritionHealthValueFromJson(
+  Map<String, dynamic> json,
+) => NutritionHealthValue(
   name: json['name'] as String?,
   mealType: json['meal_type'] as String?,
   calories: (json['calories'] as num?)?.toDouble(),
@@ -513,7 +586,9 @@ NutritionHealthValue _$NutritionHealthValueFromJson(Map<String, dynamic> json) =
   zinc: (json['zinc'] as num?)?.toDouble(),
 )..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$NutritionHealthValueToJson(NutritionHealthValue instance) => <String, dynamic>{
+Map<String, dynamic> _$NutritionHealthValueToJson(
+  NutritionHealthValue instance,
+) => <String, dynamic>{
   '__type': ?instance.$type,
   'name': ?instance.name,
   'meal_type': ?instance.mealType,
@@ -561,15 +636,42 @@ Map<String, dynamic> _$NutritionHealthValueToJson(NutritionHealthValue instance)
   'zinc': ?instance.zinc,
 };
 
-MenstruationFlowHealthValue _$MenstruationFlowHealthValueFromJson(Map<String, dynamic> json) =>
-    MenstruationFlowHealthValue(
-      flow: $enumDecodeNullable(_$MenstrualFlowEnumMap, json['flow']),
-      dateTime: DateTime.parse(json['dateTime'] as String),
-      isStartOfCycle: json['isStartOfCycle'] as bool?,
-      wasUserEntered: json['wasUserEntered'] as bool?,
-    )..$type = json['__type'] as String?;
+ActivityIntensityHealthValue _$ActivityIntensityHealthValueFromJson(
+  Map<String, dynamic> json,
+) => ActivityIntensityHealthValue(
+  intensityLevel: $enumDecode(
+    _$ActivityIntensityLevelEnumMap,
+    json['intensityLevel'],
+  ),
+  minutes: (json['minutes'] as num).toDouble(),
+)..$type = json['__type'] as String?;
 
-Map<String, dynamic> _$MenstruationFlowHealthValueToJson(MenstruationFlowHealthValue instance) => <String, dynamic>{
+Map<String, dynamic> _$ActivityIntensityHealthValueToJson(
+  ActivityIntensityHealthValue instance,
+) => <String, dynamic>{
+  '__type': ?instance.$type,
+  'intensityLevel': _$ActivityIntensityLevelEnumMap[instance.intensityLevel]!,
+  'minutes': instance.minutes,
+};
+
+const _$ActivityIntensityLevelEnumMap = {
+  ActivityIntensityLevel.moderate: 'moderate',
+  ActivityIntensityLevel.vigorous: 'vigorous',
+  ActivityIntensityLevel.unknown: 'unknown',
+};
+
+MenstruationFlowHealthValue _$MenstruationFlowHealthValueFromJson(
+  Map<String, dynamic> json,
+) => MenstruationFlowHealthValue(
+  flow: $enumDecodeNullable(_$MenstrualFlowEnumMap, json['flow']),
+  dateTime: DateTime.parse(json['dateTime'] as String),
+  isStartOfCycle: json['isStartOfCycle'] as bool?,
+  wasUserEntered: json['wasUserEntered'] as bool?,
+)..$type = json['__type'] as String?;
+
+Map<String, dynamic> _$MenstruationFlowHealthValueToJson(
+  MenstruationFlowHealthValue instance,
+) => <String, dynamic>{
   '__type': ?instance.$type,
   'flow': ?_$MenstrualFlowEnumMap[instance.flow],
   'isStartOfCycle': ?instance.isStartOfCycle,
@@ -586,16 +688,18 @@ const _$MenstrualFlowEnumMap = {
   MenstrualFlow.spotting: 'spotting',
 };
 
-WorkoutSummary _$WorkoutSummaryFromJson(Map<String, dynamic> json) => WorkoutSummary(
-  workoutType: json['workoutType'] as String,
-  totalDistance: json['totalDistance'] as num,
-  totalEnergyBurned: json['totalEnergyBurned'] as num,
-  totalSteps: json['totalSteps'] as num,
-);
+WorkoutSummary _$WorkoutSummaryFromJson(Map<String, dynamic> json) =>
+    WorkoutSummary(
+      workoutType: json['workoutType'] as String,
+      totalDistance: json['totalDistance'] as num,
+      totalEnergyBurned: json['totalEnergyBurned'] as num,
+      totalSteps: json['totalSteps'] as num,
+    );
 
-Map<String, dynamic> _$WorkoutSummaryToJson(WorkoutSummary instance) => <String, dynamic>{
-  'workoutType': instance.workoutType,
-  'totalDistance': instance.totalDistance,
-  'totalEnergyBurned': instance.totalEnergyBurned,
-  'totalSteps': instance.totalSteps,
-};
+Map<String, dynamic> _$WorkoutSummaryToJson(WorkoutSummary instance) =>
+    <String, dynamic>{
+      'workoutType': instance.workoutType,
+      'totalDistance': instance.totalDistance,
+      'totalEnergyBurned': instance.totalEnergyBurned,
+      'totalSteps': instance.totalSteps,
+    };